# Avoid OrderedIterator when comparing OpenApiSchema.Required equality